order save changes added

+++ src/main/java/lk/ijse/hello_shoes_shop_backend/entity/OrderEntity.java
@@ -24,7 +24,7 @@ public class OrderEntity implements SuperEntity{
     private int qty;
     private double total;
     private String paymentMethod;
-    private double points;
+    private int points;
     private String cashierName;
 
     @ManyToOne
